Added Security Interface and renamed Security class to SecurityManager
